Support streaming inputs with delimited protobufs.
When used with certain inputs that receive stream data (such as TCP),
the default implementation of protobuf parsing sometimes fails silently
leading to hard to debug data loss.

This commit adds a 'streaming' configuration parameter. If set to true,
data received from a logstash input is buffered until a protobuf message
can be decoded successfully.

In this setup, protobuf messages should be sent to logstash in delimited
format, that is, each serialized message should be prefixed by a varint
containing the total size of the serialized message.

For reference, here's a link to the logstash line codec which does
something similar, but looking for newlines:
https://github.com/logstash-plugins/logstash-codec-line/blob/master/lib/logstash/codecs/line.rb#L38